app.ts is a reserved file name imo - rename constants file
