+++ src/gridGL/pixiApp/PixiApp.ts
@@ -1,26 +1,26 @@
-import { Renderer, Container, Graphics } from 'pixi.js';
 import { Viewport } from 'pixi-viewport';
-import { PixiAppSettings } from './PixiAppSettings';
-import { Pointer } from '../interaction/pointer/Pointer';
-import { Update } from './Update';
-import './pixiApp.css';
-import { GridLines } from '../UI/GridLines';
-import { AxesLines } from '../UI/AxesLines';
-import { GridHeadings } from '../UI/gridHeadings/GridHeadings';
-import { Cursor } from '../UI/Cursor';
-import { Cells } from '../UI/cells/Cells';
-import { zoomInOut, zoomToFit, zoomToSelection } from '../helpers/zoom';
-import { Quadrants } from '../quadrants/Quadrants';
-import { QUADRANT_SCALE } from '../quadrants/quadrantConstants';
-import { debugAlwaysShowCache, debugNeverShowCache, debugShowCacheFlag } from '../../debugFlags';
-import { Sheet } from '../../grid/sheet/Sheet';
-import { SheetController } from '../../grid/controller/sheetController';
-import { HEADING_SIZE } from '../../constants/gridConstants';
+import { Container, Graphics, Renderer } from 'pixi.js';
 import { editorInteractionStateDefault } from '../../atoms/editorInteractionStateAtom';
 import { gridInteractionStateDefault } from '../../atoms/gridInteractionStateAtom';
-import { IS_READONLY_MODE } from '../../constants/app';
-import { Wheel, ZOOM_KEY, HORIZONTAL_SCROLL_KEY } from '../pixiOverride/Wheel';
+import { IS_READONLY_MODE } from '../../constants/appConstants';
+import { HEADING_SIZE } from '../../constants/gridConstants';
+import { debugAlwaysShowCache, debugNeverShowCache, debugShowCacheFlag } from '../../debugFlags';
+import { SheetController } from '../../grid/controller/sheetController';
+import { Sheet } from '../../grid/sheet/Sheet';
+import { AxesLines } from '../UI/AxesLines';
+import { Cursor } from '../UI/Cursor';
+import { GridLines } from '../UI/GridLines';
 import { BoxCells } from '../UI/boxCells';
+import { Cells } from '../UI/cells/Cells';
+import { GridHeadings } from '../UI/gridHeadings/GridHeadings';
+import { zoomInOut, zoomToFit, zoomToSelection } from '../helpers/zoom';
+import { Pointer } from '../interaction/pointer/Pointer';
+import { HORIZONTAL_SCROLL_KEY, Wheel, ZOOM_KEY } from '../pixiOverride/Wheel';
+import { Quadrants } from '../quadrants/Quadrants';
+import { QUADRANT_SCALE } from '../quadrants/quadrantConstants';
+import { PixiAppSettings } from './PixiAppSettings';
+import { Update } from './Update';
+import './pixiApp.css';
 
 export class PixiApp {
   private parent?: HTMLDivElement;
